backend/requirements.txt
asgiref==3.8.1
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.7
Django==5.0.6
django-ckeditor==6.7.1
django-cors-headers==4.3.1
django-js-asset==2.2.0
django-rest-knox==4.2.0
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
inflection==0.5.1
mysqlclient==2.2.4
packaging==24.0
pdfminer.six==20231228
pdfplumber==0.11.0
pillow==10.3.0
pycparser==2.22
PyJWT==2.8.0
pypdfium2==4.30.0
pytz==2024.1
PyYAML==6.0.1
setuptools==69.5.1
sqlparse==0.5.0
tzdata==2024.1
uritemplate==4.1.1
